requirements.txt
django~=5.2.1
easydict~=1.13
numpy~=1.26.4
opencv-python~=4.9.0.80
tensorflow~=2.16.1
vidgear~=0.3.2
ultralytics~=8.3.119
tqdm~=4.66.4
djangorestframework~=3.16.0
matplotlib~=3.8.4
lapx~=0.5.11
yt_dlp~=2025.3.31
channels~=4.2.0
channels-redis~=4.2.0
asgiref~=3.8.1
psycopg
geojson
opencv-python-headless
gunicorn
daphne
